# Improve the performance of fetching subscriptions by product where the query is limited to customers or orders