--- a/includes/class-wc-subscription-product-query-controller.php
+++ b/includes/class-wc-subscription-product-query-controller.php
@@ -61,10 +61,10 @@ class WC_Subscription_Product_Query_Controller {
 
 		// If we're querying by order ID or customer ID, we can filter the results by product ID after the query has been run.
 		if ( isset( $this->query_vars['order_id'] ) || isset( $this->query_vars['customer_id'] ) ) {
-			$can_filter_results = true;
+			$can_filter_results = apply_filters( 'wcs_should_filter_subscriptions_results_by_product_id', true, $this->query_vars );
 		}
 
-		return apply_filters( 'wcs_can_filter_subscriptions_results_by_product_id', $can_filter_results, $this->query_vars );
+		return $can_filter_results;
 	}
 
 	/**