--- a/wcs-functions.php
+++ b/wcs-functions.php
@@ -512,11 +512,11 @@ function wcs_get_subscriptions( $args ) {
 	// It's more efficient to filter the results by product ID or variation ID rather than querying for via a "post__in" clause.
 	// This can only work where we know that the results will be sufficiently limited by the other query args. ie when we're querying by customer_id or order_id.
 	// We store the filters in a separate array so that we can apply them after the query has been run.
-	$product_query_handler = new WC_Subscription_Product_Query_Controller( $args );
+	$query_controller = new WC_Subscription_Query_Controller( $args );
 
 	// We need to restrict subscriptions to those which contain a certain product/variation.
-	if ( $product_query_handler->has_product_query() ) {
-		if ( $product_query_handler->should_filter_query_results() ) {
+	if ( $query_controller->has_product_query() ) {
+		if ( $query_controller->should_filter_query_results() ) {
 			// We will filter the results and apply any paging, limit and offset after the query has been run.
 			unset( $args['product_id'], $args['variation_id'], $query_args['limit'], $query_args['paged'], $query_args['offset'] );
 
@@ -546,9 +546,9 @@ function wcs_get_subscriptions( $args ) {
 	}
 
 	// If we didn't query the database for subscriptions to a product, filter the results now.
-	if ( $product_query_handler->has_product_query() && $product_query_handler->should_filter_query_results() ) {
-		$subscriptions = $product_query_handler->filter_subscriptions( $subscriptions );
-		$subscriptions = $product_query_handler->paginate_results( $subscriptions );
+	if ( $query_controller->has_product_query() && $query_controller->should_filter_query_results() ) {
+		$subscriptions = $query_controller->filter_subscriptions( $subscriptions );
+		$subscriptions = $query_controller->paginate_results( $subscriptions );
 	}
 
 	return apply_filters( 'woocommerce_got_subscriptions', $subscriptions, $args );