@james-allan 2024-10-22
@mattallan I'm wondering if this approach should be limited to large sites. 

In my testing, this approach is slower on smaller sites like mine which has 1200 orders and 350 subscriptions. Having said that, it's marginal on smaller sites because we're talking 1 millisecond.

I'm going to try set up a large site to see if I can benchmark this better.