@james-allan 2024-10-22
On a hosted store (not local) running 4000 subscriptions so far (we're still in the process of creating a large site for testing), this change is leading to a marginal improvement. 

When fetching subscriptions by order ID and product, the run time of `wcs_get_subscriptions()` goes from 40.558637 to 7.641439 milliseconds. 

Because of the way the changes in this PR work, I suspect that 7 millisecond time will stay pretty stationary but the upper limit will get larger as the store gets larger.


I'll check in again once the store is larger.