@james-allan 2024-10-28
>  i.e. WC_Subscription_Query_Controller, so that we can build on it for other subscription query optimisations and not have it limited to products?

Yeah I like that idea. [9319c27](https://github.com/Automattic/woocommerce-subscriptions-core/pull/705/commits/9319c27c04f7facf9946839f703dbdec1b154c2c)


> I'm not sure if this data is skewed somehow but I'm surprised that having HPOS disabled was performing better across the board 🤔 Curious @james-allan if you noticed something similar on your test site?

I'm seeing something similar but only marginal differences. 

| Orders + Subscriptions | Posts new ms | HPOS new ms | diff (ms) | Old approach |
|--------|--------|--------|--------|--------|
| 11635 | 5.287ms | 17.294 | -12.175 | **P:** 100.405ms **HPOS:** 109.328ms | 


So in my testing HPOS is 12ms slower however HPOS is also slower on the old approach too. ie the old approach with Posts took 100.405ms and on HPOS it took 109.328ms (so hpos was also ~10ms slower).

I'm not sure why that would be.