@mattallan 2024-11-12
Thanks James for working on these improvements. I've done a bit more testing on this branch and looking forward to get this into the next release.

Merging this